Catch spot dupes if they come in very quickly
Basically (apparently) there was a hole in the PC[16]1 handling where
it would accept a duplicate spot if it (physically) came in the same
second.

I'm not completely convinced that this is root of the problem but I can't
see a reason to have the code that was there in DXProt::handle_11.